refactor: delete ErrorBoundary lib, create own ErrorBoundary component
