+++ package.json
@@ -13,7 +13,6 @@
   "dependencies": {
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-error-boundary": "^4.0.11",
     "react-router-dom": "^6.17.0"
   },
   "devDependencies": {
